requirements.txt
opencv-python-headless
numpy
scikit-learn
torch
torchvision
Pillow
louis
gradio
matplotlib
